requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision & Processing
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart
pydantic>=2.0.0
websockets>=11.0.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
pyyaml
tqdm
requests
